backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-jose[cryptography]>=3.3.0
pydantic>=2.10.0
supabase>=2.11.0
python-dotenv>=1.0.1
python-multipart>=0.0.20
apscheduler>=3.10.4
slowapi>=0.1.9
httpx>=0.27.0
email-validator>=2.0.0
